Web interface and predefined badges are working
The predefined badges interface is ugly. It's not really clear
where it should sit amongst everything, so it's just been put
into the server package.

Most likely the "external data fetching" part of the system should
go into it's own API, and appropriate DTOs to sanitize the templates
filter up to that. But it works for now.